% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_it}
\alias{map_it}
\title{map it.}
\usage{
map_it(sf, palette = "OrRd", provider = leaflet::providers$CartoDB.Positron,
  ...)
}
\arguments{
\item{sf}{An object of class `sf`}
}
\description{
maps a census variable.
}
\details{
variables need to be renamed to value instead of estimate.
}
